+++ apps/jotai-form-example/src/hooks/usePokemonQuery.ts
@@ -1,0 +1,6 @@
+import { useQuery } from '@tanstack/react-query'
+import { pokemonData } from 'mock-data'
+
+export function usePokemonQuery() {
+  return useQuery(['allPokemon'], () => pokemonData)
+}