+++ packages/apollo-components/src/DataTable/types.ts
@@ -37,12 +37,14 @@ export interface RowConfig<T> {
 
 export interface CellConfig<T> {
   getStickyCellColor?: (cell: Cell<T, unknown>) => string
+  getShouldHighlight?: (cell: Cell<T, unknown>) => boolean
 }
 
 export type RowSelectionMode = 'single' | 'multiple'
 
 export type DataTableConfig<T> = {
   height?: string
+  tableLayout?: 'auto' | 'fixed'
   sortable?: boolean
   virtual?: boolean
   rowSelectionMode?: RowSelectionMode